Deepak -- intravital -- stable -- InvivoCytometer, DNADamageAnalysis
>> Modified all GUI tools of Invivo Cytometer and DNADamageAnalyzer.m to accept more file formats

For this i write a general class called BioImageReader.m and a function called uiGetBioImageData.m

Experimenting with size features in ComputeRegionMergingFeatures.m

>> DNA Damage Analysis

DNADamageAnalyzer.m now has been extended to 

-- uses a machine learning model to detect foci -- trainFociDetectionModel.m, ComputeFociDetectionFeatures.m
-- segment the drug and macrophage channels and perform some colocalization analysis - segmentDrugCisplatin.m, segmentMacrophages.m

>> Added a new function for thresholding by removing the background using morphological opening - thresholdSBR.m

This works better on varied datasets than the local minimum error thresholding i was using for nuclei segmentation. So i plan to make this the default thresholding algorithm in the future.

>> Updated trainCellStateClassificationModel_3D.m to generate G1SG2_M model as well

>> Added some utility functions -- imresizetotarget.m


git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/intravital/main@11519 0bcb941b-74d1-0310-bb26-e6636707f451

+++ CorrectTwoPhotonConfocalStageShift.m
@@ -46,7 +46,7 @@ function [mergedImageData, imRegValidMask, varargout] = CorrectTwoPhotonConfocal
             
             % setup fixed and moving image data
             dataFixed = struct;
-            dataFixed.im = confocalImageData{1,i};
+            dataFixed.im = double( confocalImageData{1,i} );
             dataFixed.Spacing = confocalPixelSpacing;   
 
             dataMoving = struct;
@@ -93,7 +93,7 @@ function [mergedImageData, imRegValidMask, varargout] = CorrectTwoPhotonConfocal
             curRegTimer = tic;
             
             % setup fixed and moving image data
-            dataFixed.im = confocalImageData{1,i};
+            dataFixed.im = double( confocalImageData{1,i} );
             dataFixed.Spacing = confocalPixelSpacing;   
 
             dataMoving.im = imTwoPhoton;
@@ -190,7 +190,7 @@ end
 function [ imPreprocessed ] = PreprocessData( imInput )
 
     % standardize the intensity range
-    ImageIntensityRange = ComputeImageDynamicRange( imInput, 98.0 );
+    ImageIntensityRange = ComputeImageDynamicRange(double(imInput), 98.0);
     imPreprocessed = mat2gray(imInput, ImageIntensityRange) * 4096;
     
 %     % denoise it
